requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.0.1=pypi_0
anyio=3.7.0=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backoff=2.2.1=pypi_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pypi_0
blinker=1.6.2=pypi_0
bzip2=1.0.8=h3422bc3_4
ca-certificates=2023.5.7=hf0a4a13_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
chromadb=0.3.26=pypi_0
click=8.1.3=pypi_0
clickhouse-connect=0.6.2=pypi_0
cohere=4.7.0=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.1.3=pyhd8ed1ab_0
cryptography=41.0.1=pypi_0
dataclasses-json=0.5.7=pypi_0
debugpy=1.6.7=py311ha397e9f_0
decorator=5.1.1=pyhd8ed1ab_0
duckdb=0.8.0=pypi_0
executing=1.2.0=pyhd8ed1ab_0
fastapi=0.96.1=pypi_0
filelock=3.12.1=pypi_0
flatbuffers=23.5.26=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.6.0=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.19.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.2=pypi_0
grpcio-tools=1.54.2=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
h5py=3.8.0=pypi_0
hnswlib=0.7.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=0.17.2=pypi_0
httptools=0.5.0=pypi_0
httpx=0.24.1=pypi_0
huggingface-hub=0.15.1=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
ipykernel=6.23.1=pyh736e0ef_0
ipython=8.14.0=pyhd1c38e8_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
jsonschema=4.17.3=pypi_0
jupyter_client=8.2.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py311h267d04e_0
keras=2.13.1rc0=pypi_0
langchain=0.0.196=pypi_0
langchainplus-sdk=0.0.8=pypi_0
libclang=16.0.0=pypi_0
libcxx=16.0.4=h4653b0c_0
libexpat=2.5.0=hb7217d7_1
libffi=3.4.2=h3422bc3_5
libsodium=1.0.18=h27ca646_1
libsqlite=3.42.0=hb31c410_0
libzlib=1.2.13=h03a7124_4
lz4=4.3.2=pypi_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.19.0=pypi_0
marshmallow-enum=1.5.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
numexpr=2.8.4=pypi_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.15.0=pypi_0
openai=0.27.7=pypi_0
openapi-schema-pydantic=1.2.4=pypi_0
openssl=3.1.1=h53f4e23_1
opt-einsum=3.3.0=pypi_0
overrides=7.3.1=pypi_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.2=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pdfminer-six=20221105=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.1.2=pyhd8ed1ab_0
platformdirs=3.5.1=pyhd8ed1ab_0
portalocker=2.7.0=pypi_0
posthog=3.0.1=pypi_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=4.23.2=pypi_0
psutil=5.9.5=py311he2be06e_0
ptyprocess=0.7.0=pyhd3deb0d_0
pulsar-client=3.2.0=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=12.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pycryptodome=3.18.0=pypi_0
pydantic=1.10.8=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pympler=1.0.1=pypi_0
pypdf=3.9.1=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.11.3=h1456518_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pypi_0
python_abi=3.11=3_cp311
pytz=2023.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.1.0=py311hb1af645_0
readline=8.2=h92ec313_1
regex=2023.6.3=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.4.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.1=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlalchemy=2.0.15=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.27.0=pypi_0
streamlit=1.23.1=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorflow-estimator=2.13.0rc0=pypi_0
tensorflow-macos=2.13.0rc1=pypi_0
termcolor=2.3.0=pypi_0
tk=8.6.12=he1e0b03_0
tokenizers=0.13.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.3.2=py311heffc1b2_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.30.1=pypi_0
typing-extensions=4.5.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=4.3=pypi_0
urllib3=1.26.16=pypi_0
uvicorn=0.22.0=pypi_0
uvloop=0.17.0=pypi_0
validators=0.20.0=pypi_0
watchfiles=0.19.0=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
websockets=11.0.3=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
wikipedia=1.4.0=pypi_0
wrapt=1.15.0=pypi_0
xz=5.2.6=h57fd34a_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.15.0=pyhd8ed1ab_0
zstandard=0.21.0=pypi_0
